+++ sqla_utils/types.py
@@ -1,7 +1,9 @@
 """Type definitions for sqla-utils."""
 
-from typing import TypeAlias
+from typing import Any, TypeAlias, TypeVar
 
 from sqlalchemy.engine.row import Row
 
-RowType: TypeAlias = Row
+_TP = TypeVar("_TP", bound=tuple[Any, ...])
+
+RowType: TypeAlias = Row[_TP]